service-workers/service-worker/waiting.https.html

<!DOCTYPE html>
<title>ServiceWorker: navigator.serviceWorker.waiting</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="resources/test-helpers.sub.js"></script>
<body>
<script>

const SCRIPT = 'resources/empty-worker.js';
const SCOPE = 'resources/blank.html';
let registration;

promise_test(async t => {
    let frame;
    await service_worker_unregister(t, SCOPE);
    frame = await with_iframe(SCOPE);
    registration =
        await navigator.serviceWorker.register(SCRIPT, {scope: SCOPE});
    await wait_for_state(t, registration.installing, 'installed');
    const controller =
        await frame.contentWindow.navigator.serviceWorker.controller;

    assert_equals(controller, null, '.controller');
    assert_equals(registration.active, null, '.active');
    assert_equals(registration.waiting.state, 'installed', '.waiting');
    assert_equals(registration.installing, null, '.installing');
    t.add_cleanup(() => frame.remove());
    await registration.unregister();
}, 'waiting is set after installation');

// Tests that the ServiceWorker objects returned from waiting attribute getter
// that represent the same service worker are the same objects.
promise_test(async t => {
    registration =
        await service_worker_unregister_and_register(t, SCRIPT, SCOPE);
    await registration.waiting;
    const r = await navigator.serviceWorker.getRegistration(SCOPE);
    assert_equals(r.scriptURl, registration.scriptURl, 'getRegistration ' +
                  'should return the same object');
    t.add_cleanup(() => r.unregister());
}, 'The ServiceWorker objects returned from waiting attribute getter that ' +
   'represent the same service worker are the same objects');
</script>
</body>
